requirements.txt
Flask
gunicorn
google-auth 
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-storage
google-cloud-bigquery
python-dotenv
requests